refactor: transfer profile icons
