+++ src/assets/ico/MinusIcon.js
