+++ src/assets/ico/PlusIcon.js
